+++ R/p_attributes.R
@@ -8,10 +8,18 @@ NULL
 #' will have a set of positional attributes (such as part-of-speech, or lemma).
 #' The available p-attributes are returned by the \code{p_attributes}-method.
 #' 
-#' @param .Object A length-one \code{character} vector, or a \code{partition} object.
-#' @param ... further arguments
+#' The \code{p_attributes}-method returns the p-attributes defined for the
+#' corpus the partition is derived from, if argument \code{p_attribute} is
+#' \code{NULL} (the default). If \code{p_attribute} is defined, the unique
+#' values for the p-attribute are returned.
+#' 
+#' @param .Object A length-one \code{character} vector, or a \code{partition}
+#'   object.
+#' @param ... Arguments passed to \code{get_token_stream}.
 #' @param p_attribute A p-attribute to decode, provided by a length-one
 #'   \code{character} vector.
+#' @param decode A length-one \code{logical} value. Whether to return decoded
+#'   p-attributes or unique token ids.
 #' @exportMethod p_attributes
 #' @rdname p_attributes
 #' @name p_attributes
@@ -24,56 +32,75 @@ NULL
 setGeneric("p_attributes", function(.Object, ...) standardGeneric("p_attributes"))
 
 #' @rdname p_attributes
-setMethod("p_attributes", "character", function(.Object, p_attribute = NULL, ...){
-  p_attributes(.Object = corpus(.Object), p_attribute = p_attribute, ...)
+setMethod("p_attributes", "character", function(.Object, p_attribute = NULL){
+  p_attributes(.Object = corpus(.Object), p_attribute = p_attribute)
 })
 
 
 #' @rdname p_attributes
-setMethod("p_attributes", "corpus", function(.Object, p_attribute = NULL, ...){
-  if ("pAttribute" %in% names(list(...))) p_attribute <- list(...)[["pAttribute"]]
+setMethod("p_attributes", "corpus", function(.Object, p_attribute = NULL){
   p_attrs <- registry_get_p_attributes(corpus = .Object@corpus)
   if (is.null(p_attribute)){
-    return( p_attrs )
+    return(p_attrs)
   } else {
-    if (p_attribute %in% p_attrs){
-      tokens <- get_token_stream(.Object = .Object, p_attribute = p_attribute, ...)
-      return(tokens)
-    } else {
-      stop("p_attribute provided is not available")
+    if (!p_attribute %in% p_attrs){
+      stop(sprintf("The p-attribute '' is not available in corpus ''.", p_attribute, .Object@corpus))
     }
+    lexfile <- file.path(.Object@data_dir, sprintf("%s.lexicon", p_attribute), fsep = "/")
+    lexicon <- readBin(con = lexfile, what = character(), n = file.info(lexfile)$size)
+    if (.Object@encoding != localeToCharset()){
+      lexicon <- stringi::stri_encode(lexicon, from = .Object@encoding, to = localeToCharset())
+    }
+    return(lexicon)
   }
 })
 
 
-#' @details The \code{p_attributes}-method returns the p-attributes defined for the
-#' corpus the partition is derived from, if argument \code{p_attribute} is \code{NULL}
-#' (the default). If \code{p_attribute} is defined, the unique values for the p-attribute
-#' are returned.
-#' @rdname partition_class
-setMethod("p_attributes", "slice", function(.Object, p_attribute = NULL, ...){
-  if ("pAttribute" %in% names(list(...))) p_attribute <- list(...)[["pAttribute"]]
+#' @rdname p_attributes
+setMethod("p_attributes", "slice", function(.Object, p_attribute = NULL, decode = TRUE){
   p_attrs <- registry_get_p_attributes(.Object@corpus)
   if (is.null(p_attribute)){
     return( p_attrs )
   } else {
-    if (p_attribute %in% p_attrs){
-      if (p_attribute %in% .Object@p_attribute && length(p_attribute) == 1L){
-        return(.Object@stat[[p_attribute]])
-      } else {
-        return(unique(get_token_stream(.Object)))
-      }
-    } else {
-      stop("p-attribute provided is not available")
+    if (!p_attribute %in% p_attrs){
+      stop(sprintf("The p-attribute '' is not available in corpus ''.", p_attribute, .Object@corpus))
     }
+    ids <- RcppCWB::cl_cpos2id(
+      corpus = .Object@corpus,
+      p_attribute = p_attribute,
+      registry = registry,
+      cpos = cpos(.Object@cpos)
+    )
+    ids_unique <- unique(ids)
+    ids_unique <- ids_unique[order(ids_unique)]
+    str <- cl_id2str(corpus = .Object@corpus, p_attribute = p_attribute, registry = registry(), id = ids_unique)
+    if (corpus@encoding != localeToCharset()){
+      str <- stringi::stri_encode(str, from = corpus@encoding, to = localeToCharset())
+    }
+    return(str)
+  }
+})
+
+#' @rdname p_attributes
+setMethod("p_attributes", "partition_bundle", function(.Object, p_attribute = NULL, decode = TRUE){
+  corpus_id <- unique(sapply(.Object@objects, slot, "corpus"))
+  if (length(corpus_id) > 1L) stop("Getting p-attributes for a corpus requires that objects ",
+                                   "are derived from the same corpus.")
+  p_attrs <- registry_get_p_attributes(.Object@corpus)
+  if (is.null(p_attribute)){
+    return(p_attrs)
+  } else {
+    y_pre <- get_token_stream(.Object, p_attribute = p_attribute, decode = decode)
+    y <- lapply(y_pre, unique)
+    return(y)
   }
 })
 
 #' @rdname partition_class
-setMethod("p_attributes", "partition", function(.Object, p_attribute = NULL, ...) callNextMethod())
+setMethod("p_attributes", "partition", function(.Object, p_attribute = NULL, decode = TRUE) callNextMethod())
 
 #' @rdname partition_class
-setMethod("p_attributes", "subcorpus", function(.Object, p_attribute = NULL, ...) callNextMethod())
+setMethod("p_attributes", "subcorpus", function(.Object, p_attribute = NULL, decode = TRUE) callNextMethod())
 
 
 #' @rdname context-class